refactor(pod): split pina_pod_primitives into multi-file module structure
Split the monolithic lib.rs (2322 lines) into focused module files:
- pod_bool.rs: PodBool type and conversions
- pod_numeric.rs: PodU16..PodI128 via macros
- macros.rs: define_pod_unsigned!, define_pod_signed!, impl_pod_common!, etc.
- error.rs: PodCollectionError enum
- option.rs: PodOption<T> with kani proofs
- string.rs: PodString<N, PFX> with kani proofs
- vec.rs: PodVec<T, N, PFX> with kani proofs
- tests/: unit test modules for each type

All 253 tests pass (112 unit + 141 fuzz).
